Fis issue Modify block sidebar UX #131
