+++ app/image-map-query-block/editor/default-view.module.scss
