+++ app/image-map-query-block/editor/popup-panels.module.scss
@@ -21,7 +21,7 @@
 	.include {
 		font-size: 12px;
 		font-weight: 500;
-		padding-top: 1em;
+		padding-top: 24px;
 		border-top: 1px solid #e0e0e0;
 	}
